+++ src/main/java/com/aibe/team2/domain/auth/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package com.aibe.team2.domain.auth.repository;
 
-import com.aibe.team2.domain.auth.util.RefreshToken;
+import com.aibe.team2.domain.auth.entity.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
