+++ src/main/java/com/aibe/team2/domain/auth/service/CustomMemberDetailService.java
@@ -1,5 +1,6 @@
 package com.aibe.team2.domain.auth.service;
 
+import com.aibe.team2.domain.auth.dto.CustomUserDetails;
 import com.aibe.team2.domain.mypage.entity.Member;
 import com.aibe.team2.domain.mypage.repository.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
