+++ src/main/java/com/aibe/team2/domain/auth/service/CustomOAuth2UserService.java
@@ -2,7 +2,7 @@ package com.aibe.team2.domain.auth.service;
 
 import com.aibe.team2.domain.auth.entity.SocialAuth;
 import com.aibe.team2.domain.auth.repository.SocialAuthRepository;
-import com.aibe.team2.domain.auth.util.CustomUserDetails;
+import com.aibe.team2.domain.auth.dto.CustomUserDetails;
 import com.aibe.team2.domain.mypage.entity.Member;
 import com.aibe.team2.domain.mypage.entity.enums.Provider;
 import com.aibe.team2.domain.mypage.entity.enums.Role;
