+++ src/main/java/com/aibe/team2/global/common/resolver/LoginMemberIdResolver.java
@@ -1,5 +1,6 @@
 package com.aibe.team2.global.common.resolver;
 
+import com.aibe.team2.domain.auth.dto.CustomUserDetails;
 import com.aibe.team2.global.common.annotation.LoginMemberId;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,8 +38,8 @@ public class LoginMemberIdResolver implements HandlerMethodArgumentResolver {
 
         // 1. 정상적인 인증 토큰이 있는 경우 (가장 이상적)
         if (authentication != null && authentication.isAuthenticated() && authentication.getPrincipal() instanceof CustomUserDetails userDetails) {
-            if (userDetails.member() != null) {
-                return userDetails.member().getMemberId();
+            if (userDetails.getMember() != null) {
+                return userDetails.getMember().getMemberId();
             }
         }
